+++ packages/client/src/features/postModal/index.ts
@@ -1,0 +1,1 @@
+export { default as PostModal } from './ui/PostModal';
